feat(api): migrate list campaign to bounded context
