+++ api/lib/application/organizations/index.js
@@ -388,24 +388,6 @@ const register = async function (server) {
     },
     {
       method: 'GET',
-      path: '/api/organizations/{id}/campaigns',
-      config: {
-        pre: [{ method: securityPreHandlers.checkUserBelongsToOrganization }],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.organizationId,
-          }),
-        },
-        handler: organizationController.findPaginatedFilteredCampaigns,
-        tags: ['api', 'organizations'],
-        notes: [
-          'Cette route est restreinte aux utilisateurs authentifiés',
-          'Elle retourne les campagnes rattachées à l’organisation.',
-        ],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/organizations/{id}/invitations',
       config: {
         pre: [
