+++ api/lib/infrastructure/repositories/campaign-report-repository.js
@@ -2,30 +2,11 @@ import { knex } from '../../../db/knex-database-connection.js';
 import { CampaignReport } from '../../domain/read-models/CampaignReport.js';
 import { TargetProfileForSpecifier } from '../../domain/read-models/campaign/TargetProfileForSpecifier.js';
 import { CampaignParticipationStatuses } from '../../domain/models/index.js';
-import { fetchPage } from '../utils/knex-utils.js';
 import { NotFoundError } from '../../domain/errors.js';
 import _ from 'lodash';
-import { filterByFullName } from '../utils/filter-utils.js';
 import * as campaignRepository from './campaign-repository.js';
 
 const { SHARED } = CampaignParticipationStatuses;
-
-function _setSearchFiltersForQueryBuilder(qb, { name, ongoing = true, ownerName, isOwnedByMe }, userId) {
-  if (name) {
-    qb.whereILike('name', `%${name}%`);
-  }
-  if (ongoing) {
-    qb.whereNull('campaigns.archivedAt');
-  } else {
-    qb.whereNotNull('campaigns.archivedAt');
-  }
-  if (ownerName) {
-    filterByFullName(qb, ownerName, 'users.firstName', 'users.lastName');
-  }
-  if (isOwnedByMe) {
-    qb.where('users.id', '=', userId);
-  }
-}
 
 const get = async function (id) {
   const result = await knex('campaigns')
@@ -112,33 +93,4 @@ const findMasteryRatesAndValidatedSkillsCount = async function (campaignId) {
   return aggregatedResults;
 };
 
-const findPaginatedFilteredByOrganizationId = async function ({ organizationId, filter, page, userId }) {
-  const query = knex('campaigns')
-    .distinct('campaigns.id')
-    .select(
-      'campaigns.*',
-      'users.id AS "ownerId"',
-      'users.firstName AS ownerFirstName',
-      'users.lastName AS ownerLastName',
-      knex.raw(
-        'COUNT(*) FILTER (WHERE "campaign-participations"."id" IS NOT NULL AND "campaign-participations"."isImproved" IS FALSE AND "campaign-participations"."deletedAt" IS NULL) OVER (partition by "campaigns"."id") AS "participationsCount"',
-      ),
-      knex.raw(
-        'COUNT(*) FILTER (WHERE "campaign-participations"."id" IS NOT NULL AND "campaign-participations"."status" = \'SHARED\' AND "campaign-participations"."isImproved" IS FALSE AND "campaign-participations"."deletedAt" IS NULL) OVER (partition by "campaigns"."id") AS "sharedParticipationsCount"',
-      ),
-    )
-    .join('users', 'users.id', 'campaigns.ownerId')
-    .leftJoin('campaign-participations', 'campaign-participations.campaignId', 'campaigns.id')
-    .where('campaigns.organizationId', organizationId)
-    .modify(_setSearchFiltersForQueryBuilder, filter, userId)
-    .orderBy('campaigns.createdAt', 'DESC');
-
-  const { results, pagination } = await fetchPage(query, page);
-  const atLeastOneCampaign = await knex('campaigns').select('id').where({ organizationId }).first(1);
-  const hasCampaigns = Boolean(atLeastOneCampaign);
-
-  const campaignReports = results.map((result) => new CampaignReport(result));
-  return { models: campaignReports, meta: { ...pagination, hasCampaigns } };
-};
-
-export { get, findMasteryRatesAndValidatedSkillsCount, findPaginatedFilteredByOrganizationId };
+export { get, findMasteryRatesAndValidatedSkillsCount };
