+++ api/tests/prescription/campaign/integration/application/campaign-administration-route_test.js
@@ -2,7 +2,7 @@ import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 import { campaignAdministrationController } from '../../../../../src/prescription/campaign/application/campaign-adminstration-controller.js';
 import * as moduleUnderTest from '../../../../../src/prescription/campaign/application/campaign-administration-route.js';
 
-describe('Integration | Application | Route | campaignRouter', function () {
+describe('Integration | Application | Route | campaign administration router', function () {
   let httpTestServer;
 
   beforeEach(async function () {
