+++ api/tests/prescription/campaign/unit/domain/usecases/find-paginated-filtered-organization-campaigns_test.js
@@ -1,5 +1,5 @@
-import { expect, domainBuilder, sinon } from '../../../test-helper.js';
-import { findPaginatedFilteredOrganizationCampaigns } from '../../../../lib/domain/usecases/find-paginated-filtered-organization-campaigns.js';
+import { expect, domainBuilder, sinon } from '../../../../../test-helper.js';
+import { findPaginatedFilteredOrganizationCampaigns } from '../../../../../../src/prescription/campaign/domain/usecases/find-paginated-filtered-organization-campaigns.js';
 
 describe('Unit | Domain | Use Cases | find-paginated-filtered-organization-campaigns', function () {
   const campaignReportRepository = { findPaginatedFilteredByOrganizationId: () => undefined };
