+++ api/tests/unit/application/organizations/organization-controller_test.js
@@ -164,106 +164,6 @@ describe('Unit | Application | Organizations | organization-controller', functio
     });
   });
 
-  describe('#findPaginatedFilteredCampaigns', function () {
-    let organizationId;
-    let request;
-    let campaign;
-    let serializedCampaigns;
-    let dependencies;
-
-    beforeEach(function () {
-      organizationId = 1;
-      request = {
-        params: { id: organizationId },
-        auth: {
-          credentials: {
-            userId: 1,
-          },
-        },
-      };
-      campaign = domainBuilder.buildCampaign();
-      serializedCampaigns = [{ name: campaign.name, code: campaign.code }];
-
-      const queryParamsUtilsStub = {
-        extractParameters: sinon.stub(),
-      };
-      const campaignReportSerializerStub = {
-        serialize: sinon.stub(),
-      };
-      dependencies = {
-        queryParamsUtils: queryParamsUtilsStub,
-        campaignReportSerializer: campaignReportSerializerStub,
-      };
-      sinon.stub(usecases, 'findPaginatedFilteredOrganizationCampaigns');
-    });
-
-    it('should call the usecase to get the campaigns and associated campaignReports', async function () {
-      // given
-      request.query = {
-        campaignReport: true,
-      };
-      const expectedPage = 2;
-      const expectedFilter = { name: 'Math' };
-      dependencies.queryParamsUtils.extractParameters
-        .withArgs(request.query)
-        .returns({ page: expectedPage, filter: expectedFilter });
-      const expectedResults = [campaign];
-      const expectedPagination = { page: expectedPage, pageSize: 25, itemsCount: 100, pagesCount: 4 };
-      usecases.findPaginatedFilteredOrganizationCampaigns.resolves({
-        models: expectedResults,
-        pagination: expectedPagination,
-      });
-      dependencies.campaignReportSerializer.serialize.returns({ data: serializedCampaigns, meta: {} });
-
-      // when
-      await organizationController.findPaginatedFilteredCampaigns(request, hFake, dependencies);
-
-      // then
-      expect(usecases.findPaginatedFilteredOrganizationCampaigns).to.have.been.calledWithExactly({
-        organizationId,
-        filter: expectedFilter,
-        page: expectedPage,
-        userId: request.auth.credentials.userId,
-      });
-    });
-
-    it('should return the serialized campaigns belonging to the organization', async function () {
-      // given
-      request.query = {};
-      const expectedResponse = { data: serializedCampaigns, meta: {} };
-      dependencies.queryParamsUtils.extractParameters.withArgs({}).returns({ filter: {} });
-      usecases.findPaginatedFilteredOrganizationCampaigns.resolves({ models: {}, pagination: {} });
-      dependencies.campaignReportSerializer.serialize.returns(expectedResponse);
-
-      // when
-      const response = await organizationController.findPaginatedFilteredCampaigns(request, hFake, dependencies);
-
-      // then
-      expect(response).to.deep.equal(expectedResponse);
-    });
-
-    it('should return a JSON API response with meta information', async function () {
-      // given
-      request.query = {};
-      const expectedResults = [campaign];
-      const expectedPagination = { page: 2, pageSize: 25, itemsCount: 100, pagesCount: 4, hasCampaigns: true };
-      dependencies.queryParamsUtils.extractParameters.withArgs({}).returns({ filter: {} });
-      usecases.findPaginatedFilteredOrganizationCampaigns.resolves({
-        models: expectedResults,
-        meta: expectedPagination,
-      });
-
-      // when
-      await organizationController.findPaginatedFilteredCampaigns(request, hFake, dependencies);
-
-      // then
-      expect(dependencies.campaignReportSerializer.serialize).to.have.been.calledWithExactly(
-        expectedResults,
-        expectedPagination,
-      );
-    });
-  });
-
   describe('#findPaginatedCampaignManagements', function () {
     let organizationId;
     let request;
